moving SWIFT message handling into a separate library
