+++ src/Buchen/Buchen.csproj
@@ -13,14 +13,8 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\libfintx.EBICS.Letters\libfintx.EBICS.Letters.csproj" />
-    <ProjectReference Include="..\libfintx.EBICS.Swift\libfintx.EBICS.Swift.csproj" />
     <ProjectReference Include="..\libfintx.EBICSConfig\libfintx.EBICSConfig.csproj" />
     <ProjectReference Include="..\libfintx.EBICS\libfintx.EBICS.csproj" />
-    <ProjectReference Include="..\libfintx.Security\libfintx.Security.csproj" />
-    <ProjectReference Include="..\libfintx.StatePrinter\libfintx.StatePrinter.csproj" />
-    <ProjectReference Include="..\libfintx.Xml\libfintx.Xml.csproj" />
-    <ProjectReference Include="..\libfintx.Xsd\libfintx.Xsd.csproj" />
   </ItemGroup>
 
 </Project>
