+++ src/libfintx.Swift/SwiftLine.cs
@@ -2,7 +2,8 @@
  * 	
  *  This file is part of libfintx.
  *  
- *  Copyright (C) 2021 - 2022 Abid Hussain
+ *  Copyright (C) 2016 - 2022 Torsten Klinger
+ * 	E-Mail: torsten.klinger@googlemail.com
  *  
  *  This program is free software; you can redistribute it and/or
  *  modify it under the terms of the GNU Lesser General Public
@@ -20,25 +21,17 @@
  * 	
  */
 
-namespace libfintx.FinTS.Swift
-{
-    public enum SepaPurpose
-    {
-        // https://www.hettwer-beratung.de/sepa-spezialwissen/sepa-technische-anforderungen/sepa-gesch%C3%A4ftsvorfallcodes-gvc-mt-940/
+namespace libfintx.Swift;
 
-        IBAN, // SEPA IBAN Auftraggeber
-        BIC, // SEPA BIC Auftraggeber
-        EREF, // SEPA End to End-Referenz
-        KREF, // Kundenreferenz
-        MREF, // SEPA Mandatsreferenz
-        CRED, // SEPA Creditor Identifier
-        DEBT, // Originator Identifier
-        COAM, // Zinskompensationsbetrag
-        OAMT, // Ursprünglicher Umsatzbetrag
-        SVWZ, // SEPA Verwendungszweck
-        ABWA, // Abweichender SEPA Auftraggeber
-        ABWE, // Abweichender SEPA Empfänger
-        BREF, // Bankreferenz, Instruction ID
-        RREF // Retourenreferenz
+public class SwiftLine
+{
+    public string SwiftTag { get; set; }
+
+    public string SwiftData { get; set; }
+
+    public SwiftLine(string swiftTag, string swiftData)
+    {
+        SwiftTag = swiftTag;
+        SwiftData = swiftData;
     }
 }
