+++ src/libfintx.Swift/docs/README.md
@@ -1,0 +1,18 @@
+﻿# libfintx.Swift
+[![License: LGPL v3](https://img.shields.io/badge/License-LGPL%20v3-blue.svg)](https://www.gnu.org/licenses/lgpl-3.0)
+
+This library holds functions to handle SWIFT messages.
+
+&nbsp;
+
+# Features
+
+* MT940 message parsing
+* MT942 message parsing
+
+
+# Copyright & License
+
+Licensed under **GNU LESSER GENERAL PUBLIC LICENSE Version 3, 29 June 2007**. Please read the [LICENSE](https://github.com/libfintx/libfintx/blob/main/LICENSE) file.
+
+Copyright (c) 2016 - 2024 **Torsten Klinger**
